cmd: Rename jx to jnx and split out flag version
Rename `jx` to `jnx` as the jx binary name is already taken by jenkinsx.
But jnx is also a better name since it sounds like jinx, and that's what
you say when someone duplicates a word at the same time as you. jsonnet
is about de-duplicating through templating, so jnx works. It can also be
pronounced as a single syllable, which is also nice.

Split the Go flag version of jnx into a completely separate program
rather than having a `parseCLI()` function selected by build tags.

The only reason for the flag version is to use the flag code path of
jsonnext to demonstrate its use and to verify it works. jnx itself is
going to evolve in future to contain subcommands which I do not intend
to support in the flag version of jnx. So split it out so jnx can evolve
on its own.

Keep `jx.jsonnet` for now for backwards compatibility. It will be
removed when I know I've removed all existing uses under that name.
