+++ cmd/jnx/main.go
@@ -4,20 +4,22 @@ import (
 	"fmt"
 	"os"
 
-	jxkong "foxygo.at/jsonnext/kong"
+	jnxkong "foxygo.at/jsonnext/kong"
+	"github.com/alecthomas/kong"
 	jsonnet "github.com/google/go-jsonnet"
 )
 
 type config struct {
-	jxkong.Config
+	jnxkong.Config
 	Filename string `arg:"" optional:"" help:"File to evaluate. stdin is used if omitted or \"-\""`
 }
 
 func main() {
-	cli := parseCLI()
-	vm := cli.Config.MakeVM("JXPATH")
+	c := &config{Config: *jnxkong.NewConfig()}
+	kong.Parse(c)
+	vm := c.Config.MakeVM("JNXPATH")
 
-	out, err := run(vm, cli.Filename)
+	out, err := run(vm, c.Filename)
 	if err != nil {
 		fmt.Fprintln(os.Stderr, err)
 		os.Exit(1)
