+++ lib/jnx.jsonnet
@@ -1,0 +1,9 @@
+// Package jnx is a top-level package to import the subpackages of the
+// jnx library.
+{
+  array:: import 'array.jsonnet',
+  object:: import 'object.jsonnet',
+  op:: import 'op.jsonnet',
+  string:: import 'string.jsonnet',
+  value:: import 'value.jsonnet',
+}
