+++ lib/jx.jsonnet
@@ -1,9 +1,2 @@
-// Package jx is a top-level package to import the subpackages of the
-// jx library.
-{
-  array:: import 'array.jsonnet',
-  object:: import 'object.jsonnet',
-  op:: import 'op.jsonnet',
-  string:: import 'string.jsonnet',
-  value:: import 'value.jsonnet',
-}
+// Kept for backwards compatibilty for now. Will be removed in the future.
+import 'jnx.jsonnet'
